add hexagon buffer map class
